fix(vendor): strip jsr version meta module graph data
